Remove unused `endpoints` from SDK configs
`endpoints` was added in order to dynamically add new API endpoints to
SDK. The idea was that if the API added a new endpoint, that could be
also added by the SDK's user to the SDK, without the need to upgrade SDK
version. However, this was never really used and the current SDK code
doesn't read the `endpoints` configuration.