Add loginAs
And deprecate login

We'll add an explicit loginAs function. This way, we can avoid checking
the grantType to define if we are using loginAs functionality. Grant
type is only the mechanism we use to authenticate and we might use the
same authorization code mechanism to something else in the future.