Make requireAuth less confusing
If authentication fails, requireAuth returns the return value of
reject() callback, which is `undefined`. However, we use the return
value as a Promise chain and call `.then()` to it. Obviously,
`undefined.then()` should fail. However, that failure is never visible
due to the fact that we are already inside a rejected Promise context,
which will catch the exception and return the original rejection reason.