Deduplicate auth requests
When multiple API calls are made simultaneusly, SDK will fire multiple
auth token API calls. That's wasteful, and this commit changes the
implementation so that we do only one auth API call and the other API
calls wait for that one call to complete before proceeding.

This is done by storing the auth API call promise to a mutable
"in-flight request store", which is essentially just a normal mutable JS
object.